Clean up application styles
