+++ App/Sources/UI/Views/Extensions/View+Extensions.swift
@@ -87,7 +87,7 @@ struct DebugView<Content>: View where Content: View {
       .padding()
     }
     .buttonStyle(
-      GradientButtonStyle(.init(nsColor: .systemIndigo))
+      AppButtonStyle(.init(nsColor: .systemIndigo))
     )
     .background(
       LinearGradient(colors: [
