+++ App/Sources/UI/Views/NewCommand/NewCommandShortcutView.swift
@@ -33,7 +33,7 @@ struct NewCommandShortcutView: View {
       }
       .background(NewCommandValidationView($validation))
     }
-    .menuStyle(GradientMenuStyle(.init(nsColor: .systemGray), fixedSize: false))
+    .menuStyle(AppMenuStyle(.init(nsColor: .systemGray), fixedSize: false))
     .onChange(of: validation, perform: { newValue in
       guard newValue == .needsValidation else { return }
       withAnimation { validation = updateAndValidatePayload() }
