+++ App/Sources/UI/Views/WorkflowCommandListHeaderView.swift
@@ -41,7 +41,7 @@ struct WorkflowCommandListHeaderView: View {
             .padding(.vertical, 4)
             .offset(x: -4, y: 1)
         })
-        .menuStyle(GradientMenuStyle(.init(nsColor: .systemGray), menuIndicator: .visible))
+        .menuStyle(AppMenuStyle(.init(nsColor: .systemGray), menuIndicator: .visible))
         .frame(maxWidth: detailPublisher.data.execution == .concurrent ? 144 : 110,
                alignment: .leading)
       }
